+++ java/shared/java/org/signal/libsignal/internal/Native.kt
@@ -1024,7 +1024,7 @@ internal object Native {
   @JvmStatic @Throws(Exception::class)
   public external fun SenderCertificate_New(senderUuid: String, senderE164: String?, senderDeviceId: Int, senderKey: ObjectHandle, expiration: Long, signerCert: ObjectHandle, signerKey: ObjectHandle): ObjectHandle
   @JvmStatic @Throws(Exception::class)
-  public external fun SenderCertificate_Validate(cert: ObjectHandle, key: ObjectHandle, time: Long): Boolean
+  public external fun SenderCertificate_Validate(cert: ObjectHandle, trustRoots: LongArray, time: Long): Boolean
 
   @JvmStatic @Throws(Exception::class)
   public external fun SenderKeyDistributionMessage_Deserialize(data: ByteArray): ObjectHandle