+++ java/shared/java/org/signal/libsignal/metadata/certificate/CertificateValidator.kt
@@ -8,25 +8,41 @@ package org.signal.libsignal.metadata.certificate
 import org.signal.libsignal.internal.Native
 import org.signal.libsignal.protocol.ecc.ECPublicKey
 
+/**
+ * Used by [SealedSessionCipher](org.signal.libsignal.metadata.SealedSessionCipher) to validate
+ * sealed sender certificates.
+ */
 public open class CertificateValidator(
-  public val trustRoot: ECPublicKey,
+  public val trustRoots: List<ECPublicKey>,
 ) {
+  public constructor(trustRoot: ECPublicKey) : this(listOf(trustRoot))
+
+  /**
+   * Validates `certificate`.
+   *
+   * The default behavior checks the certificate against each key in [trustRoots] in constant time
+   * (that is, no result is produced until every key is checked), making sure **one** of them has
+   * signed its embedded server certificate.
+   * The `validationTime` parameter is compared numerically against [SenderCertificate.expiration]
+   * and is not required to use any specific units, but Signal uses milliseconds since 1970.
+   *
+   * @throws InvalidCertificateException if the certificate is invalid or has expired
+   */
   @Throws(InvalidCertificateException::class)
-  public fun validate(
+  public open fun validate(
     certificate: SenderCertificate,
     validationTime: Long,
   ) {
     try {
       certificate.guardedRun { certificateHandle ->
-        trustRoot.guardedRun { trustRootHandle ->
-          if (!Native.SenderCertificate_Validate(
-              certificateHandle,
-              trustRootHandle,
-              validationTime,
-            )
-          ) {
-            throw InvalidCertificateException("Validation failed")
-          }
+        val trustRootHandles = trustRoots.map { it.unsafeNativeHandleWithoutGuard() }.toLongArray()
+        if (!Native.SenderCertificate_Validate(
+            certificateHandle,
+            trustRootHandles,
+            validationTime,
+          )
+        ) {
+          throw InvalidCertificateException("Validation failed")
         }
       }
     } catch (e: Exception) {
@@ -35,10 +51,17 @@ public open class CertificateValidator(
   }
 
   // VisibleForTesting
+
+  /**
+   * Verifies a ServerCertificate against the set of trust roots.
+   *
+   * This is only meant for convenience in unit testing and thus is *not* a constant-time operation.
+   * It is not used by SealedSessionCipher.
+   */
   @Throws(InvalidCertificateException::class)
   public fun validate(certificate: ServerCertificate) {
-    if (!trustRoot.verifySignature(certificate.certificate, certificate.signature)) {
-      throw InvalidCertificateException("Signature failed")
+    if (!trustRoots.any { it.verifySignature(certificate.certificate, certificate.signature) }) {
+      throw InvalidCertificateException("All signatures failed")
     }
   }
 }