+++ RELEASE_NOTES.md
@@ -1,2 +1,3 @@
 v0.79.1
 
+- Java: `CertificateValidator.validate(SenderCertificate, long)` is once again `open` for testing.